+++ lib/Doctrine/ODM/PHPCR/Mapping/ClassMetadataFactory.php
@@ -133,6 +133,10 @@ class ClassMetadataFactory extends AbstractClassMetadataFactory
             $this->getDriver()->loadMetadataForClass($class->getName(), $class);
         }
 
+        if ($parent && $class->getInheritMixins()) {
+            $class->setMixins(array_merge($parent->getMixins(), $class->getMixins()));
+        }
+
         if ($this->evm->hasListeners(Event::loadClassMetadata)) {
             $eventArgs = new LoadClassMetadataEventArgs($class, $this->dm);
             $this->evm->dispatchEvent(Event::loadClassMetadata, $eventArgs);
@@ -155,7 +159,6 @@ class ClassMetadataFactory extends AbstractClassMetadataFactory
         $subClass->setVersioned($parentClass->versionable);
         $subClass->setReferenceable($parentClass->referenceable);
         $subClass->setNodeType($parentClass->getNodeType());
-        $subClass->setMixins($parentClass->getMixins());
     }
 
     /**